moving enums into a folder to cleanup autonomous folder
