+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/enums/Bar.java
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.enums;
+
+public enum Bar
+{
+  HIGH_BAR,
+  LOW_BAR
+}
