+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/enums/RotateDirection.java
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.enums;
+
+public enum RotateDirection
+{
+  RIGHT,
+  LEFT
+}
