--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modules/drive/RotateDirection.java
@@ -1,7 +1,0 @@
-package org.firstinspires.ftc.teamcode.modules.drive;
-
-public enum RotateDirection
-{
-  RIGHT,
-  LEFT
-}
