--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/Bar.java
@@ -1,7 +1,0 @@
-package org.firstinspires.ftc.teamcode.opmode;
-
-public enum Bar
-{
-  HIGH_BAR,
-  LOW_BAR
-}
