+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/Gamepads.java
@@ -2,6 +2,9 @@ package org.firstinspires.ftc.teamcode.opmode;
 
 import com.qualcomm.robotcore.hardware.Gamepad;
 
+import org.firstinspires.ftc.teamcode.enums.Button;
+import org.firstinspires.ftc.teamcode.enums.Participant;
+
 import java.util.Set;
 
 public class Gamepads
