--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/Team.java
@@ -1,7 +1,0 @@
-package org.firstinspires.ftc.teamcode.opmode;
-
-public enum Team
-{
-  RED,
-  BLUE
-}
