+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/autonomous/BlueTeamSpecimens.java
@@ -1,6 +1,8 @@
-package org.firstinspires.ftc.teamcode.opmode;
+package org.firstinspires.ftc.teamcode.opmode.autonomous;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+import org.firstinspires.ftc.teamcode.enums.Team;
 
 @Autonomous( name = "Blue Team Specimens", group = "13702" )
 public class BlueTeamSpecimens extends AbstractAutonomousOpMode
