+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/autonomous/GameStrategy.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmode;
+package org.firstinspires.ftc.teamcode.opmode.autonomous;
 
 public enum GameStrategy
 {
