+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/autonomous/RedTeamBaskets.java
@@ -1,6 +1,8 @@
-package org.firstinspires.ftc.teamcode.opmode;
+package org.firstinspires.ftc.teamcode.opmode.autonomous;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+import org.firstinspires.ftc.teamcode.enums.Team;
 
 @Autonomous( name = "Red Team Basket", group = "13702" )
 public class RedTeamBaskets extends AbstractAutonomousOpMode
